+++ src/main/kotlin/cn/spacexc/bilibilisdk/sdk/article/info/remote/content/Pendant.kt
@@ -1,10 +1,8 @@
 package cn.spacexc.bilibilisdk.sdk.article.info.remote.content
 
 
-import com.google.gson.annotations.SerializedName
-
 data class Pendant(
-    val pid: Int,
+    val pid: Long,
     val name: String,
     val image: String,
     val expire: Int
